requirements.txt
# API and Web Framework
fastapi==0.104.1
uvicorn==0.23.2

# Database
sqlalchemy==2.0.23

# Data Validation
pydantic==2.4.2

# Authentication
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# File Handling
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==1.0.0

# AI Integration
openai==1.3.0

# Configuration
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.1
asyncio==3.4.3
